refactor: Update middleware module imports and remove unused files
